HJBQVI - lbfgs bug fix, refactor
